+++ packages/annotation/src/misc/index.tsx
@@ -1,2 +1,3 @@
+export * from './Connector'
 export * from './GridStripes'
 export * from './types'